Исправлен тест CheckCurrentTsar и комментарий. Папка проекта TestsForHomeExercises переименована.
