+++ cs/TestsForHomeExercises/TestsForHomeExercises.csproj
@@ -6,10 +6,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <None Remove="C:\Users\lizas\.nuget\packages\nunit.engine\3.11.1\contentFiles\any\netstandard2.0\nunit.engine.nuget.addins" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="FluentAssertions" Version="5.10.3" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.7.1" />
     <PackageReference Include="NUnit" Version="3.12.0" />
